prototype/backend/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework
django-cors-headers
django-filter
djangorestframework-gis
